# apps/admin-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/task": {
      "get": {
        "operationId": "TaskController_getTasksByLecture",
        "parameters": [
          {
            "name": "lectureId",
            "required": true,
            "in": "query",
            "description": "조회할 과제가 속한 강의의 ID입니다.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 강의에 속한 모든 과제를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "과제를 조회하는 데 실패했습니다."
          }
        },
        "summary": "특정 강의의 과제 목록 조회",
        "tags": [
          "Task"
        ]
      },
      "post": {
        "operationId": "TaskController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새로운 과제를 생성합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEntity"
                }
              }
            }
          },
          "400": {
            "description": "과제 생성에 실패했습니다."
          }
        },
        "summary": "새로운 과제 생성",
        "tags": [
          "Task"
        ]
      }
    },
    "/task/{id}": {
      "get": {
        "operationId": "TaskController_getTaskById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 ID를 가진 과제를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEntity"
                }
              }
            }
          },
          "400": {
            "description": "과제를 조회하는 데 실패했습니다."
          },
          "404": {
            "description": "과제를 찾을 수 없습니다."
          }
        },
        "summary": "특정 과제 조회",
        "tags": [
          "Task"
        ]
      },
      "patch": {
        "operationId": "TaskController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "특정 ID를 가진 과제를 업데이트합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEntity"
                }
              }
            }
          },
          "400": {
            "description": "과제 업데이트에 실패했습니다."
          },
          "404": {
            "description": "업데이트하려는 과제를 찾을 수 없습니다."
          }
        },
        "summary": "특정 과제 수정",
        "tags": [
          "Task"
        ]
      },
      "delete": {
        "operationId": "TaskController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 ID를 가진 과제를 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEntity"
                }
              }
            }
          },
          "400": {
            "description": "과제 삭제에 실패했습니다."
          },
          "404": {
            "description": "삭제하려는 과제를 찾을 수 없습니다."
          }
        },
        "summary": "특정 과제 삭제",
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sinchon Admin Backend",
    "description": "신촌연합 관리자 페이지 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TaskEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "고유 ID 값"
          },
          "round": {
            "type": "number",
            "description": "과제가 속한 강의의 회차"
          },
          "practiceId": {
            "type": "number",
            "description": "해당 강의의 BOJ 그룹 내에서 과제를 나타내는 연습의 ID (BOJ 크롤링 시 사용)"
          },
          "minSolveCount": {
            "type": "number",
            "description": "과제 출석 여부를 판단하기 위한 기준 문제 수"
          },
          "lectureId": {
            "type": "number",
            "description": "과제가 속한 강의의 ID"
          }
        },
        "required": [
          "id",
          "round",
          "practiceId",
          "minSolveCount",
          "lectureId"
        ]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "round": {
            "type": "number",
            "description": "강의 회차 (round)"
          },
          "practiceId": {
            "type": "number",
            "description": "BOJ 그룹 내 연습 ID (practiceId)"
          },
          "lectureId": {
            "type": "number",
            "description": "연결된 강의 ID (Lecture의 id 값)"
          },
          "minSolveCount": {
            "type": "number",
            "description": "출석 여부를 판단하는 최소 문제 해결 수"
          }
        },
        "required": [
          "round",
          "practiceId",
          "lectureId",
          "minSolveCount"
        ]
      },
      "CreateProblemDto": {
        "type": "object",
        "properties": {
          "bojProblemNumber": {
            "type": "number",
            "description": "BOJ 문제 번호"
          },
          "essential": {
            "type": "boolean",
            "description": "필수 문제 여부"
          },
          "taskId": {
            "type": "number",
            "minimum": 1,
            "description": "과제 ID (Task의 id 값)"
          }
        },
        "required": [
          "bojProblemNumber",
          "essential",
          "taskId"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "round": {
            "type": "number",
            "description": "강의 회차 (round)"
          },
          "practiceId": {
            "type": "number",
            "description": "BOJ 그룹 내 연습 ID (practiceId)"
          },
          "lectureId": {
            "type": "number",
            "description": "연결된 강의 ID (Lecture의 id 값)"
          },
          "minSolveCount": {
            "type": "number",
            "description": "출석 여부를 판단하는 최소 문제 해결 수"
          },
          "problems": {
            "description": "문제 배열 (BOJ 문제 정보)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProblemDto"
            }
          }
        }
      },
...
    }
  }
}